+++ packages/element-lib/src/__tests__/PoisonedBatchFileAttack.spec.js
@@ -14,6 +14,7 @@ const wrongBatchFileHash = 'QmTJGHccriUtq3qf3bvAQUcDUHnBbHNJG2x2FYwYUecN43';
 
 beforeAll(async () => {
   sidetree = getTestSideTree();
+  sidetree.parameters.mapSync = false;
   await sidetree.db.deleteDB();
   await generateActors(sidetree, 1);
   actor = await getActorByIndex(0);